Register component has been implemented, and post and reply components upgraded.

+++ GoldenForum.Service/Models/ViewModels/Post/PostListViewModel.cs
@@ -14,7 +14,7 @@ namespace GoldenForum.Service.Models.ViewModels.Post
         public string AuthorId { get; set; }
         public string AuthorUserName { get; set; }
         public string AuthorImageUrl { get; set; }
-        public double AuthorRating { get; set; }
+        public int AuthorRating { get; set; }
 
         public int RepliesCount { get; set; }
     }
